Propagate count of available USB devices to Redux
When multiple supported devices are available for connection we may
show some additional errors to the users to have only one device active.

Devices available := Sum(Count of each paired VID/PID connected)

The LedgerJS library only supports communication with a single fixed
device, and the responding device is not known to us in advance.